BaseActivity seems to work ok.  In the next commit, we'll try doing a live update on the LimboActivity page to show the user his account status being reviewed and approved by the admin.  That should be pretty cool.
